Updates testing proj with more azure search bits

+++ src/Examine.Web.Demo/Models/TestModel.cs
@@ -22,7 +22,7 @@ namespace Examine.Web.Demo.Models
 
     public class IndexInfo
     {
-        public int Docs { get; set; }
+        public long Docs { get; set; }
         public int Fields { get; set; }
     }
